5.1.3 確実にリソースを解放する 5.1.4 自動リソース管理（ARM）の使用 5.2 ラムダ式でリソース解放 5.2.1 リソース解放を行うクラスの準備 5.2.2 高階関数の使用 5.2.3 インスタンス解放に使用 5.3 ロックの管理
