+++ Chapter05/resources/fpij/FileWriterARM_5_1_2.java
@@ -1,5 +1,3 @@
-package Chapter05.resources.jpij;
-
 import java.io.FileWriter;
 import java.io.IOException;
